pyproject.toml
[tool.poetry]
name = "petbot"
version = "0.1.0"
description = ""
authors = ["kbondar17 <ironbondar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.31"
python-dotenv = "^0.19.2"
aiogram = "^2.18"
pydantic = "^1.9.0"
psycopg2-binary = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
alembic = "^1.7.5"
typer = "^0.4.0"
google-api-python-client = "^2.36.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
httpx = "^0.22.0"
yadisk = "^1.2.15"
Faker = "^12.0.0"
aiocache = "^0.11.1"
python-telegram-bot-pagination = "^0.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
